rename OutfitEditorCategory to OutfitEditorMenu
